+++ src/pocketmine/entity/Human.php
@@ -702,7 +702,7 @@ class Human extends Creature implements ProjectileSource, InventoryHolder{
 
 			if($this->foodTickTimer === 0){
 				if($food >= 18){
-					if($health < $this->getMaxHealth() and $this->level->getGameRules()->getBool(GameRules::RULE_NATURAL_REGENERATION)){
+					if($health < $this->getMaxHealth() and $this->level->getGameRules()->getBool(GameRules::RULE_NATURAL_REGENERATION, true)){
 						$this->heal(new EntityRegainHealthEvent($this, 1, EntityRegainHealthEvent::CAUSE_SATURATION));
 						$this->exhaust(3.0, PlayerExhaustEvent::CAUSE_HEALTH_REGEN);
 					}
