+++ src/pocketmine/entity/passive/Sheep.php
@@ -77,7 +77,7 @@ class Sheep extends Animal{
 		return "Sheep";
 	}
 
-	public function onInteract(Player $player, Item $item, Vector3 $clickPos, int $slot) : bool{
+	public function onInteract(Player $player, Item $item, Vector3 $clickPos) : bool{
 		if(!$this->isImmobile()){
 			if($item instanceof Shears and !$this->isSheared()){
 				$this->setSheared(true);
@@ -104,7 +104,7 @@ class Sheep extends Animal{
 				return true;
 			}
 		}
-		return parent::onInteract($player, $item, $clickPos, $slot);
+		return parent::onInteract($player, $item, $clickPos);
 	}
 
 	public function getXpDropAmount() : int{
